requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
magic-filter==1.0.12
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
propcache==0.2.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyproject_hooks==1.2.0
python-dotenv==1.0.1
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.7.4
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
yarl==1.17.1
